requirements.txt
# Install PyTorch separately to match your CUDA version.
# Example (CUDA 11.3):
#   conda install pytorch==1.10.1 torchvision==0.11.2 torchaudio==0.10.1 cudatoolkit=11.3 -c pytorch -c conda-forge

# Core
numpy==1.21.6
scipy==1.7.3
pandas==1.3.5
tqdm==4.64.0
PyYAML==5.4.1
yacs==0.1.8

# Data processing and visualization
h5py==3.7.0
imageio==2.21.1
matplotlib==3.5.3
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-image==0.19.3

# Model / training utilities
einops==0.6.1
GitPython==3.1.31
loguru==0.6.0
spatial-correlation-sampler==0.4.0
tensorboard==2.10.1
thop==0.1.1.post2209072238
timm==0.6.12

# Optional dependencies for RVT / temporal SSM components
bbox-visualizer==0.1.0
hdf5plugin==4.1.1
hydra-core==1.3.2
lightning==2.1.4
omegaconf==2.3.0
plotly==5.18.0
pycocotools==2.0.7
pytorch-lightning==2.1.4
wandb==0.16.3
